Now Variable and VariableList are in 2 separate files.
